move cosmos only decorators to its own folder
