+++ app/ante/cosmos/min_price.go
@@ -1,0 +1,86 @@
+// Copyright 2021 Evmos Foundation
+// This file is part of Evmos' Ethermint library.
+//
+// Evmos is free software: you can redistribute it and/or modify
+// it under the terms of the GNU Lesser General Public License as published by
+// the Free Software Foundation, either version 3 of the License, or
+// (at your option) any later version.
+//
+// The Evmos packages are distributed in the hope that it will be useful,
+// but WITHOUT ANY WARRANTY; without even the implied warranty of
+// MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
+// GNU Lesser General Public License for more details.
+//
+// You should have received a copy of the GNU Lesser General Public License
+// along with the Ethermint library. If not, see https://github.com/evmos/evmos/blob/main/LICENSE
+package cosmos
+
+import (
+	errorsmod "cosmossdk.io/errors"
+	sdk "github.com/cosmos/cosmos-sdk/types"
+	errortypes "github.com/cosmos/cosmos-sdk/types/errors"
+	"math/big"
+)
+
+// MinGasPriceDecorator will check if the transaction's fee is at least as large
+// as the MinGasPrices param. If fee is too low, decorator returns error and tx
+// is rejected. This applies for both CheckTx and DeliverTx
+// If fee is high enough, then call next AnteHandler
+// CONTRACT: Tx must implement FeeTx to use MinGasPriceDecorator
+type MinGasPriceDecorator struct {
+	feesKeeper FeeMarketKeeper
+	evmKeeper  EVMKeeper
+}
+
+// NewMinGasPriceDecorator creates a new MinGasPriceDecorator instance used only for
+// Cosmos transactions.
+func NewMinGasPriceDecorator(fk FeeMarketKeeper, ek EVMKeeper) MinGasPriceDecorator {
+	return MinGasPriceDecorator{feesKeeper: fk, evmKeeper: ek}
+}
+
+func (mpd MinGasPriceDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
+	feeTx, ok := tx.(sdk.FeeTx)
+	if !ok {
+		return ctx, errorsmod.Wrapf(errortypes.ErrInvalidType, "invalid transaction type %T, expected sdk.FeeTx", tx)
+	}
+
+	minGasPrice := mpd.feesKeeper.GetParams(ctx).MinGasPrice
+
+	// Short-circuit if min gas price is 0 or if simulating
+	if minGasPrice.IsZero() || simulate {
+		return next(ctx, tx, simulate)
+	}
+	evmParams := mpd.evmKeeper.GetParams(ctx)
+	evmDenom := evmParams.GetEvmDenom()
+	minGasPrices := sdk.DecCoins{
+		{
+			Denom:  evmDenom,
+			Amount: minGasPrice,
+		},
+	}
+
+	feeCoins := feeTx.GetFee()
+	gas := feeTx.GetGas()
+
+	requiredFees := make(sdk.Coins, 0)
+
+	// Determine the required fees by multiplying each required minimum gas
+	// price by the gas limit, where fee = ceil(minGasPrice * gasLimit).
+	gasLimit := sdk.NewDecFromBigInt(new(big.Int).SetUint64(gas))
+
+	for _, gp := range minGasPrices {
+		fee := gp.Amount.Mul(gasLimit).Ceil().RoundInt()
+		if fee.IsPositive() {
+			requiredFees = requiredFees.Add(sdk.Coin{Denom: gp.Denom, Amount: fee})
+		}
+	}
+
+	if !feeCoins.IsAnyGTE(requiredFees) {
+		return ctx, errorsmod.Wrapf(errortypes.ErrInsufficientFee,
+			"provided fee < minimum global fee (%s < %s). Please increase the gas price.",
+			feeCoins,
+			requiredFees)
+	}
+
+	return next(ctx, tx, simulate)
+}
