+++ app/ante/evm/fees.go
@@ -26,16 +26,6 @@ import (
 	evmtypes "github.com/evmos/ethermint/x/evm/types"
 )
 
-// MinGasPriceDecorator will check if the transaction's fee is at least as large
-// as the MinGasPrices param. If fee is too low, decorator returns error and tx
-// is rejected. This applies for both CheckTx and DeliverTx
-// If fee is high enough, then call next AnteHandler
-// CONTRACT: Tx must implement FeeTx to use MinGasPriceDecorator
-type MinGasPriceDecorator struct {
-	feesKeeper FeeMarketKeeper
-	evmKeeper  EVMKeeper
-}
-
 // EthMinGasPriceDecorator will check if the transaction's fee is at least as large
 // as the MinGasPrices param. If fee is too low, decorator returns error and tx
 // is rejected. This applies to both CheckTx and DeliverTx and regardless
@@ -56,12 +46,6 @@ type EthMempoolFeeDecorator struct {
 	evmKeeper EVMKeeper
 }
 
-// NewMinGasPriceDecorator creates a new MinGasPriceDecorator instance used only for
-// Cosmos transactions.
-func NewMinGasPriceDecorator(fk FeeMarketKeeper, ek EVMKeeper) MinGasPriceDecorator {
-	return MinGasPriceDecorator{feesKeeper: fk, evmKeeper: ek}
-}
-
 // NewEthMinGasPriceDecorator creates a new MinGasPriceDecorator instance used only for
 // Ethereum transactions.
 func NewEthMinGasPriceDecorator(fk FeeMarketKeeper, ek EVMKeeper) EthMinGasPriceDecorator {
@@ -74,53 +58,6 @@ func NewEthMempoolFeeDecorator(ek EVMKeeper) EthMempoolFeeDecorator {
 	return EthMempoolFeeDecorator{
 		evmKeeper: ek,
 	}
-}
-
-func (mpd MinGasPriceDecorator) AnteHandle(ctx sdk.Context, tx sdk.Tx, simulate bool, next sdk.AnteHandler) (newCtx sdk.Context, err error) {
-	feeTx, ok := tx.(sdk.FeeTx)
-	if !ok {
-		return ctx, errorsmod.Wrapf(errortypes.ErrInvalidType, "invalid transaction type %T, expected sdk.FeeTx", tx)
-	}
-
-	minGasPrice := mpd.feesKeeper.GetParams(ctx).MinGasPrice
-
-	// Short-circuit if min gas price is 0 or if simulating
-	if minGasPrice.IsZero() || simulate {
-		return next(ctx, tx, simulate)
-	}
-	evmParams := mpd.evmKeeper.GetParams(ctx)
-	evmDenom := evmParams.GetEvmDenom()
-	minGasPrices := sdk.DecCoins{
-		{
-			Denom:  evmDenom,
-			Amount: minGasPrice,
-		},
-	}
-
-	feeCoins := feeTx.GetFee()
-	gas := feeTx.GetGas()
-
-	requiredFees := make(sdk.Coins, 0)
-
-	// Determine the required fees by multiplying each required minimum gas
-	// price by the gas limit, where fee = ceil(minGasPrice * gasLimit).
-	gasLimit := sdk.NewDecFromBigInt(new(big.Int).SetUint64(gas))
-
-	for _, gp := range minGasPrices {
-		fee := gp.Amount.Mul(gasLimit).Ceil().RoundInt()
-		if fee.IsPositive() {
-			requiredFees = requiredFees.Add(sdk.Coin{Denom: gp.Denom, Amount: fee})
-		}
-	}
-
-	if !feeCoins.IsAnyGTE(requiredFees) {
-		return ctx, errorsmod.Wrapf(errortypes.ErrInsufficientFee,
-			"provided fee < minimum global fee (%s < %s). Please increase the gas price.",
-			feeCoins,
-			requiredFees)
-	}
-
-	return next(ctx, tx, simulate)
 }
 
 // AnteHandle ensures that the that the effective fee from the transaction is greater than the
